add full Lesson7
