+++ Lesson7/src/main/java/HomeWork.java
@@ -2,18 +2,18 @@ public class HomeWork {
     //1) задача!
     /*
     Необходимо написать военкомат.
-    1) Создаем класс военкомат MilitaryOffice
-    2) Создать в классе MilitaryOffice конструктор, который принимает PersonRegistry
-    3) PersonRegistry содержит список людей (Person), которые могут быть призваны в армию
-    4) Создать класс Person, который будет содержать следующие поля имя, возраст, пол, адрес(страна, город)
+    1) Создаем класс военкомат by.tms.model.MilitaryOffice
+    2) Создать в классе by.tms.model.MilitaryOffice конструктор, который принимает by.tms.model.PersonRegistry
+    3) by.tms.model.PersonRegistry содержит список людей (by.tms.model.Person), которые могут быть призваны в армию
+    4) Создать класс by.tms.model.Person, который будет содержать следующие поля имя, возраст, пол, адрес(страна, город)
     Военкомат должен уметь:
-    a) вывести имена всех людей годных к военной службе на текущий момент времени, которые есть в PersonRegistry
-    T.е нужно написать метод в классе MilitaryOffice, который будет отфильтровывать годных к службе призывников
+    a) вывести имена всех людей годных к военной службе на текущий момент времени, которые есть в by.tms.model.PersonRegistry
+    T.е нужно написать метод в классе by.tms.model.MilitaryOffice, который будет отфильтровывать годных к службе призывников
     (годными считать мущин от 18 до 27 лет, проверяем пол и возраст)
     б) вывести количество годных призывников в городе Минске.
     в) вывести количество призывников от 25 до 27 лет
     г) вывести количество призывников у которых имя Александр.
-    5) Создать класс Main, наполнить PersonRegistry людьми, по желанию можно создавать людей через консоль
+    5) Создать класс Main, наполнить by.tms.model.PersonRegistry людьми, по желанию можно создавать людей через консоль
     */
 
     //2) задача!
