+++ Lesson7/src/main/java/by/tms/model/Car.java
@@ -5,14 +5,13 @@ public class Car {
     private String color; //можно потом добавить
     private String bodyType; //можно потом добавить
     private int yearOfRelease; //можно потом добавить
-    private final String engineType; //обязательное, нельзя изменить
     private final GasTank gasTank; //обязательное, нельзя изменить
-    private final Engine engine = new Engine(false);
+    private final Engine engine;
     private Odometer odometer = new Odometer(0);
 
     public Car(String engineType, int gasTankVolume) {
-        this.engineType = engineType;
         this.gasTank = new GasTank(gasTankVolume);
+        this.engine = new Engine(engineType);
     }
 
     public void setYearOfRelease(int yearOfRelease) {
@@ -74,8 +73,7 @@ public class Car {
     public void driving() {
         if (engine.isOn() && gasTank.getVolume() >= 5) { //если машина заведена и есть топливо на поездку, то поехали
             System.out.println("*Машина поехала!*");
-            gasTank.burningFuel(gasTank,5);
-            odometer.trip();
+            gasTank.burningFuel(gasTank,5); //сжигаем из бака 5 литров
             System.out.println("*Мы приехали!*"); //приехали
             System.out.println();
             if (gasTank.getVolume() == 0) { //если по окончанию поездки топливо на нуле, машина глохнет
@@ -83,6 +81,7 @@ public class Car {
                 System.out.println("*Машина заглохла: кончилось топливо*");
                 System.out.println();
             }
+            odometer.trip();
         } else { // не заведена и нет топлива на поездку
             System.out.println("Ошибка! Машина не заведена или не хватает топлива на поездку!");
             System.out.println();
@@ -95,85 +94,3 @@ public class Car {
 
 }
 
-class Engine {
-    private boolean isOn;
-
-    public Engine(boolean isOn) {
-        this.isOn = isOn;
-    }
-
-    public boolean isOn() {
-        return isOn;
-    }
-
-    public void start(GasTank gasTank, Engine engine) {
-
-        if (!engine.isOn) {
-            if (gasTank.getVolume() > 0) {
-                engine.isOn = true;
-                System.out.println("*Двигатель запущен*");
-                System.out.println();
-            } else {
-                System.out.println("Двигатель не запущен: нет топлива");
-                System.out.println();
-            }
-        } else {
-            System.out.println("Ошибка! Двигатель уже запущен!");
-            System.out.println();
-        }
-    }
-
-    public void off() {
-        if (isOn) {
-            isOn = false;
-            System.out.println("*Двигатель выключен*");
-            System.out.println();
-        } else {
-            System.out.println("Ошибка! Двигатель уже выключен!");
-            System.out.println();
-        }
-    }
-}
-
-class GasTank {
-
-    private int fullVolume;
-    private int volume;
-
-    public GasTank(int fullVolume) {
-        this.fullVolume = fullVolume;
-    }
-
-    public int getVolume() {
-        return volume;
-    }
-
-    public void setVolume(int volume) {
-        this.volume = volume;
-    }
-
-    public int getFullVolume() {
-        return fullVolume;
-    }
-
-    public void burningFuel(GasTank gasTank, int volume) {
-        gasTank.volume -= volume;
-    }
-}
-
-class Odometer { //прибор для измерения пробега автомобиля
-    private int fullDistance;
-
-    public Odometer(int fullDistance) {
-        this.fullDistance = fullDistance;
-    }
-
-    public int getFullDistance() {
-        return fullDistance;
-    }
-
-    public void trip() {
-        fullDistance += 10;
-    }
-
-}
